application/requirements.txt
boto3==1.26.154
botocore==1.29.154
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
comtypes==1.2.0
dnspython==2.3.0
docx2pdf==0.1.8
idna==3.4
jmespath==1.0.1
keyboard==0.13.5
Pillow==9.5.0
psutil==5.9.5
pymongo==4.3.3
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==306
pywinauto==0.6.8
reportlab==4.0.4
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
tkhtmlview==0.2.0
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
